+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org)
 * A possible segmentation fault when resizing the terminal.
 * Screens/item_details: correctly account for fields spanning multiple lines.
 * TUI: possible blank screen until first key input.
+* Screens/index: utilize 100% of terminal width.
 
 ### [v0.7.0] - 2019-02-28